Move preview code to a separate file
